+++ BetterChests/BetterChests.cs
@@ -57,6 +57,7 @@ public class BetterChests : Mod
             new CraftFromChest(this.Config, this.Helper, this.Services),
             new CustomColorPicker(this.Config, this.Helper, this.Services),
             new FilterItems(this.Config, this.Helper, this.Services),
+            new MenuForShippingBin(this.Config, this.Helper, this.Services),
             new OpenHeldChest(this.Config, this.Helper, this.Services),
             new ResizeChest(this.Config, this.Helper, this.Services),
             new ResizeChestMenu(this.Config, this.Helper, this.Services),