+++ BetterChests/Features/Feature.cs
@@ -79,6 +79,7 @@ internal abstract class Feature : IModService
             CraftFromChest => this.Config.DefaultChest.CraftFromChest != FeatureOptionRange.Disabled,
             CustomColorPicker => this.Config.DefaultChest.CustomColorPicker != FeatureOption.Disabled,
             FilterItems => this.Config.DefaultChest.FilterItems != FeatureOption.Disabled,
+            MenuForShippingBin => this.Config.CategorizeChest || this.Config.DefaultChest.ChestMenuTabs != FeatureOption.Disabled || this.Config.DefaultChest.ResizeChestMenu != FeatureOption.Disabled || this.Config.DefaultChest.SearchItems != FeatureOption.Disabled,
             OpenHeldChest => this.Config.DefaultChest.OpenHeldChest != FeatureOption.Disabled,
             ResizeChest => this.Config.DefaultChest.ResizeChest != FeatureOption.Disabled,
             ResizeChestMenu => this.Config.DefaultChest.ResizeChestMenu != FeatureOption.Disabled,