+++ EasyAccess/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Easy Access Change Log
 
+## 1.0.0-beta.2 (Unreleased)
+
+### Changes
+
+* Purge inaccessible cached objects.
+
 ## 1.0.0-beta.1 (February 12, 2022)
 
 * Initial Version